move "function_traits.h" from fluid to phi
